requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
distro==1.8.0
efficientnet-pytorch==0.6.3
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fiona==1.9.5
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.0
geopandas==0.14.1
gradio==4.7.1
gradio-client==0.7.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langchain==0.0.345
langchain-core==0.0.9
langsmith==0.0.69
lazy-loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
munch==4.0.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
openai==1.3.7
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
orjson==3.9.10
osm2gmns==0.7.3
osmium==3.7.0
osmnx==1.8.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pretrainedmodels==0.7.4
pydantic==2.5.2
pydantic-core==2.14.5
pydub==0.25.1
pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.31.1
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
segmentation-models-pytorch==0.2.0
semantic-version==2.10.0
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.4.12
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typer==0.9.0
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==11.0.3
xpinyin==0.7.6
yarl==1.9.3
